# docs/openapi/codewithmee-v1.openapi.json
{
  "components": {
    "schemas": {
...
      "Membership": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "joinedAt": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "organizationId": {
            "format": "uuid",
            "type": "string"
          },
          "role": {
            "enum": [
              "admin",
              "analyst",
              "grader",
              "instructor",
              "owner"
            ],
            "type": "string"
          },
          "status": {
            "enum": [
              "active",
              "revoked",
              "suspended"
            ],
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/MembershipUser"
          }
        },
        "required": [
          "id",
          "joinedAt",
          "organizationId",
          "role",
          "status",
          "user"
        ],
        "type": "object"
      },
...
      "Organization": {
        "additionalProperties": false,
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "maxLength": 2000,
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "industry": {
            "maxLength": 100,
            "type": "string"
          },
          "logoFile": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "maxLength": 120,
            "minLength": 2,
            "type": "string"
          },
          "ownerUserId": {
            "format": "uuid",
            "type": "string"
          },
          "revision": {
            "minimum": 1,
            "type": "integer"
          },
          "slug": {
            "pattern": "^[a-z0-9](?:[a-z0-9-]{1,61}[a-z0-9])$",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "verificationStatus": {
            "enum": [
              "approved",
              "draft",
              "pending_review",
              "rejected",
              "suspended"
            ],
            "type": "string"
          }
        },
        "required": [
          "description",
          "id",
          "industry",
          "logoFile",
          "name",
          "slug",
          "verificationStatus"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "csrfHeader": {
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      },
      "refreshCookie": {
        "in": "cookie",
        "name": "cwm_refresh",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Implemented CodeWithMee v1 operations. Unknown request fields are rejected. Future roadmap operations are intentionally omitted until deployed.",
    "title": "CodeWithMee API",
    "version": "1.0.0"
  },
...
  "openapi": "3.1.1",
  "paths": {
...
    "/organizations": {
      "get": {
        "operationId": "listMyOrganizations",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "organizations": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "membership": {
                            "$ref": "#/components/schemas/Membership"
                          },
                          "organization": {
                            "$ref": "#/components/schemas/Organization"
                          }
                        },
                        "required": [
                          "membership",
                          "organization"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "organizations"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response."
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "A stable problem-details response."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List active organization memberships",
        "tags": [
          "Organizations"
        ],
        "x-codewithmee-contract": "listMyOrganizations"
      },
      "post": {
        "operationId": "createOrganization",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "membership": {
                      "$ref": "#/components/schemas/Membership"
                    },
                    "organization": {
                      "$ref": "#/components/schemas/Organization"
                    }
                  },
                  "required": [
                    "membership",
                    "organization"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response."
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "A stable problem-details response."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create an organization with an owner membership",
        "tags": [
          "Organizations"
        ],
        "x-codewithmee-contract": "createOrganization",
        "parameters": [
          {
            "in": "header",
            "name": "idempotency-key",
            "required": false,
            "schema": {
              "maxLength": 128,
              "minLength": 16,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{15,127}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "description": {
                    "maxLength": 2000,
                    "type": "string"
                  },
                  "industry": {
                    "maxLength": 100,
                    "type": "string"
                  },
                  "name": {
                    "maxLength": 120,
                    "minLength": 2,
                    "type": "string"
                  },
                  "slug": {
                    "pattern": "^[a-z0-9](?:[a-z0-9-]{1,61}[a-z0-9])$",
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "slug"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        }
      }
    },
    "/organization-invitations/{token}/accept": {
      "post": {
        "operationId": "acceptOrganizationInvitation",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "invitation": {
                      "$ref": "#/components/schemas/Invitation"
                    },
                    "membership": {
                      "$ref": "#/components/schemas/Membership"
                    }
                  },
                  "required": [
                    "invitation",
                    "membership"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response."
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "A stable problem-details response."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Accept an organization invitation",
        "tags": [
          "Organizations"
        ],
        "x-codewithmee-contract": "acceptOrganizationInvitation",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "maxLength": 512,
              "minLength": 20,
              "pattern": "^oi1\\.",
              "type": "string"
            }
          }
        ]
      }
    },
    "/admin/provider-verifications": {
      "get": {
        "operationId": "listProviderVerificationReviews",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "reviews": {
                      "items": {
                        "$ref": "#/components/schemas/VerificationReview"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "reviews"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response."
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "A stable problem-details response."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List provider verification reviews",
        "tags": [
          "Organizations",
          "Administration"
        ],
        "x-codewithmee-contract": "listProviderVerificationReviews",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "",
                "approved",
                "pending_review",
                "rejected"
              ],
              "type": "string"
            }
          }
        ]
      }
    },
    "/admin/provider-verifications/{reviewId}/decision": {
      "post": {
        "operationId": "decideProviderVerification",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "organization": {
                      "$ref": "#/components/schemas/Organization"
                    },
                    "review": {
                      "$ref": "#/components/schemas/VerificationReview"
                    }
                  },
                  "required": [
                    "organization",
                    "review"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response."
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "A stable problem-details response."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Approve or reject a provider verification",
        "tags": [
          "Organizations",
          "Administration"
        ],
        "x-codewithmee-contract": "decideProviderVerification",
        "parameters": [
          {
            "in": "path",
            "name": "reviewId",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,99}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "reason": {
                    "maxLength": 2000,
                    "type": "string"
                  },
                  "status": {
                    "enum": [
                      "approved",
                      "rejected"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "status"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        }
      }
    },
    "/organizations/{organizationId}": {
      "get": {
        "operationId": "getOrganization",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "organization": {
                      "$ref": "#/components/schemas/Organization"
                    }
                  },
                  "required": [
                    "organization"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response."
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "A stable problem-details response."
          }
        },
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get a public or authorized private organization",
        "tags": [
          "Organizations"
        ],
        "x-codewithmee-contract": "getOrganization",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,99}$",
              "type": "string"
            }
          }
        ]
      },
      "patch": {
        "operationId": "updateOrganization",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "organization": {
                      "$ref": "#/components/schemas/Organization"
                    }
                  },
                  "required": [
                    "organization"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response."
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "A stable problem-details response."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update organization fields with optimistic concurrency",
        "tags": [
          "Organizations"
        ],
        "x-codewithmee-contract": "updateOrganization",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,99}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "description": {
                    "maxLength": 2000,
                    "type": "string"
                  },
                  "industry": {
                    "maxLength": 100,
                    "type": "string"
                  },
                  "name": {
                    "maxLength": 120,
                    "minLength": 2,
                    "type": "string"
                  },
                  "revision": {
                    "minimum": 1,
                    "type": "integer"
                  }
                },
                "required": [
                  "revision"
                ],
                "type": "object",
                "minProperties": 2
              }
            }
          },
          "required": true
        }
      }
    },
    "/organizations/{organizationId}/members": {
      "get": {
        "operationId": "listOrganizationMembers",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "members": {
                      "items": {
                        "$ref": "#/components/schemas/Membership"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "members"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response."
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "A stable problem-details response."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List authorized organization members",
        "tags": [
          "Organizations"
        ],
        "x-codewithmee-contract": "listOrganizationMembers",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,99}$",
              "type": "string"
            }
          }
        ]
      }
    },
    "/organizations/{organizationId}/invitations": {
      "post": {
        "operationId": "inviteOrganizationMember",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "invitation": {
                      "$ref": "#/components/schemas/Invitation"
                    }
                  },
                  "required": [
                    "invitation"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response."
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "A stable problem-details response."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Invite a member into an organization role",
        "tags": [
          "Organizations"
        ],
        "x-codewithmee-contract": "inviteOrganizationMember",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,99}$",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "idempotency-key",
            "required": false,
            "schema": {
              "maxLength": 128,
              "minLength": 16,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{15,127}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "email": {
                    "format": "email",
                    "type": "string"
                  },
                  "role": {
                    "enum": [
                      "admin",
                      "analyst",
                      "grader",
                      "instructor"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "role"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        }
      }
    },
    "/organizations/{organizationId}/members/{userId}": {
      "patch": {
        "operationId": "updateOrganizationMember",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "membership": {
                      "$ref": "#/components/schemas/Membership"
                    }
                  },
                  "required": [
                    "membership"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response."
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "A stable problem-details response."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update an organization membership",
        "tags": [
          "Organizations"
        ],
        "x-codewithmee-contract": "updateOrganizationMember",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,99}$",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,99}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "role": {
                    "enum": [
                      "admin",
                      "analyst",
                      "grader",
                      "instructor"
                    ],
                    "type": "string"
                  },
                  "status": {
                    "enum": [
                      "active",
                      "revoked",
                      "suspended"
                    ],
                    "type": "string"
                  }
                },
                "type": "object",
                "minProperties": 1
              }
            }
          },
          "required": true
        }
      },
      "delete": {
        "operationId": "removeOrganizationMember",
        "responses": {
          "204": {
            "description": "Successful response with no representation."
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "A stable problem-details response."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Revoke an organization membership",
        "tags": [
          "Organizations"
        ],
        "x-codewithmee-contract": "removeOrganizationMember",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,99}$",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,99}$",
              "type": "string"
            }
          }
        ]
      }
    },
    "/organizations/{organizationId}/verification": {
      "post": {
        "operationId": "submitProviderVerification",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "organization": {
                      "$ref": "#/components/schemas/Organization"
                    },
                    "review": {
                      "$ref": "#/components/schemas/VerificationReview"
                    }
                  },
                  "required": [
                    "organization",
                    "review"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful response."
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "A stable problem-details response."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Submit an organization for provider verification",
        "tags": [
          "Organizations"
        ],
        "x-codewithmee-contract": "submitProviderVerification",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,99}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "statement": {
                    "maxLength": 2000,
                    "minLength": 20,
                    "type": "string"
                  }
                },
                "required": [
                  "statement"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
}